Revert "Endpoints post"
