--- src/connection.ts
@@ -1,18 +1,0 @@
-import knex from "knex";
-import dotenv from "dotenv";
-
-dotenv.config();
-
-const connection = knex({
-  client: "mysql",
-  connection: {
-    host: process.env.DB_HOST,
-    port: Number(process.env.DB_PORT || "3306"),
-    user: process.env.DB_USER,
-    password: process.env.DB_PASSWORD,
-    database: process.env.DB_SCHEMA,
-		multipleStatements: true
-  },
-});
-
-export default connection
