--- src/data/selectAgeById.ts
@@ -1,6 +1,0 @@
-import connection from "../connection";
-
-export default async function selectAgeById(
-    id:number) {
-    const result = await connection('students')
-}
