--- src/endpoints/addStudent copy.ts
@@ -1,14 +1,0 @@
-import { Request, Response } from "express"
-
-
-export default async function addStudents(
-   req: Request,
-   res: Response
-): Promise<void> {
-
-   try {
-
-   } catch (error) {
-      res.status(500).end()
-   }
-}
