+++ src/index.ts
@@ -1,15 +1,32 @@
-import app from "./app"
-import addExpertises from "./endpoints/addExpertises"
-import addStudent from "./endpoints/addExpertises"
-import createClass from "./endpoints/createClass"
-import createStudent from "./endpoints/createStudent"
-import createTeacher from "./endpoints/createTeacher"
-import getAge from "./endpoints/getAge"
-import teacherExpertise from "./endpoints/teacherExpertise"
+import express, {Express} from 'express'
+import cors from 'cors'
+import { AddressInfo } from "net";
+import knex from "knex";
+import dotenv from "dotenv";
 
-app.put("/especialidade", addExpertises);
-app.post("/teacher-expertise", teacherExpertise);
-app.put("/create-student", createStudent);
-app.put("/create-teacher", createTeacher);
-app.post("/create-class", createClass);
-app.get("/get-age/:id", getAge);
+dotenv.config();
+
+export const connection = knex({
+	client: "mysql",
+	connection: {
+    host: process.env.DB_HOST,
+    port: 3306,
+    user: process.env.DB_USER,
+    password: process.env.DB_PASS,
+    database: process.env.DB_NAME
+  }
+});
+
+const app: Express = express();
+
+app.use(express.json());
+app.use(cors());
+
+const server = app.listen(process.env.PORT || 3003, () => {
+    if (server) {
+       const address = server.address() as AddressInfo;
+       console.log(`Server is running in http://localhost: ${address.port}`);
+    } else {
+       console.error(`Failure upon starting server.`);
+    }
+});
